[logdbg-ext] Minimal amount of configuration for debugger
Think this is close to the minimal amount of configuration needed to start up the
DAP inline in the extension. Unfortunately the extension docs are a little too detailed
around all the various ways to get the plug in started, but wish there was just one
offered and they'd go into the protocol more.