[DebugAdapter] Add support for stepping
Can step forward and backwards between lines. Assumes the initial state will start on
the first line of the file. On a stacktraceRequest, will compare the first breakpoint
with the first line. If the breakpoint is farther from the current state of line, then
will use it. Otherwise, will stop on the line. This will probably need some re-work once
think through how to represent the "debugger" between the source and log.