[DebugAdapter] Implement continue and reverseContinue
Add continue and reverseContinue methods. These will seek the next or previous breakpoint or
stop at the beginning or last line of a file. This "cheats" a little and uses the `wc` utility
to find the lines of the log file. Will need to revisit when get to reading logs from stdin or
multiple log files.