[DebugAdapter] Fixed adding breakpoints and refactored stackTraceRequest
The setBreakpoints method receives breakpoints as an argument, so using that
to figure out what breakpoints are set. This fixes a bug where the breakpoint
array would store breakpoints out of order, which didn't work when trying to locate
the next valid breakpoint.

Also modified how the stackTraceRequest such that it uses the current line as
the reference to construct the stack trace from.