[lib] Start wiring up a call graph that can query for all stack frames
Started building out a call graph that can be used to find all the stack frames
that may exist on the path from the source ref's function. At the moment will
only return the functions as nodes, but next will add edges for each call expression
within a function. eg. if there ia a function named `main` and a call expression
called `init` is declared, then there will be an edge main -> init. From this we
could find the path from main to any edge in the graph to try and determine the
stack frames that exist between for a certain line of source code.