[lib] [wip] More querying to discover the call graph for a source file
For the toy CallGraph there will exist an edge in the graph from any function (x) to another (y)
if x contains a call to y. This misses some details that would be needed to fully create a stack, but
will work for the toy prototype that is in the works.