[lib] [wip] Add SourceRef to the CallGraph's Edges
In order to re-create some of the stack, the graph should be annotated with the SourceRef for each invocation. That way
when building the stack, all the accompanying SourceRef are already present. Then in order to build the stack the
algorithm can find a path from "main" to whereever the current location in the source code is... or maybe vice versa.