[wip] [lib] Continuing to experiment with stack
Added a function that will probably find all possible paths from main to some
target function. Kinda hack-y and untested, but works with the basic examples that
exist at the moment.

Since it returns _all_ possible paths then there may be cases where there are multiple
paths that may represent the stack. In those cases, I wasn't to sure how to infer which
would be the correct one. I'd kind of like to contain all the logic outside of the
debug adapter so that the tool can remain independent, but not sure that's going to
be easy without introducing some kind of state within it... but to figure out later.