[Debug] Add syntax highlighting to log when using extension
The way I'd originally configured the debugger was to use a separate language-id, but one downside
was to remove the nice syntax highlighting that the default Log language provided. I tried for a bit
to figure out how to extend that Log lanugage with the debugger, but whenever I did I could no longer
set breakpoints.

As a workaround I ended up copy-pasting the syntax highlighting from the VS Code built-in extension and
use it in this project. I don't love the copy-pasta, but until I can figure out how to properly get the
break point UI working when using a Log language, I'm not sure what to do.