[Lib][LogRef] Use LogFormat to isolate body when extracting vars
In order to match a variable with the log line a regex is used. Previously this was using
\w, which isn't correct since if there was punctuation, then the capture would stop at it.
This tries to improve the variable extraction by using the log format to isolate the body and then
using the more permissive .+ to capture the variables.

There's some additional refactoring I'd like to do to combine LogRef and LogFormat into a singular struct
now that there is more there.