[discovery] add a layer to keep track of source tree state
Since lnav is a long-running process, we need to be able to
rescan a source tree without repeating a lot of the same
work.  This change adds the source_hier module that is a
layer that keeps track of the contents of the source
hierarchy.  It can detect changes in the files/directories
so the caller can be more targeted in figuring out what
work to do.  This new module replaces the parts of
code_source that did similar work and now CodeSource is
mostly used as an intermediate.  Since the CodeSource
objects are no longer retained, the problem of keeping all
of the source in memory is fixed as well.

Files:
* Cargo.toml: Add itertools for chunks() and some test helpers
* code_source.rs: Remove functions that are obsoleted by
  source_hier.  Add CodeSource::from_string() method for
  testing.
* lib.rs: Use source_hier in place of code_source.